+++ htdocs/custom/bookticket/passenger_card.php
@@ -205,7 +205,7 @@ if ($action == 'update' && $usercancreate)
 	}
 }
 
-// Action clone object
+/* Action clone object
 if ($action == 'confirm_clone' && $confirm != 'yes') { $action = ''; }
 if ($action == 'confirm_clone' && $confirm == 'yes' && $usercancreate)
 {
@@ -267,11 +267,11 @@ if ($action == 'confirm_clone' && $confirm == 'yes' && $usercancreate)
 			dol_print_error($db, $object->error);
 		}
 	}
-}
+}*/
 
 // Delete a passenger
-if ($action == 'confirm_delete' && $confirm != 'yes') { $action = ''; }
-if ($action == 'confirm_delete' && $confirm == 'yes' && $usercandelete)
+//if ($action == 'confirm_delete' && $confirm != 'yes') { $action = ''; }
+if ($action == 'delete' && $usercandelete)
 {
 	$result = $object->delete($user);
 
@@ -671,7 +671,7 @@ if (is_object($objcanvas) && $objcanvas->displayCanvasExists($action)) {
 	}
 }
 
-$formconfirm = '';
+/*$formconfirm = '';
 
 // Confirm delete Passenger
 if (($action == 'delete' && (empty($conf->use_javascript_ajax) || !empty($conf->dol_use_jmobile)))	// Output when action = clone if jmobile or no js
@@ -697,7 +697,7 @@ if (($action == 'clone' && (empty($conf->use_javascript_ajax) || !empty($conf->d
 
 
 // Print form confirm
-print $formconfirm;
+print $formconfirm;*/
 
 /* ************************************************************************** */
 /*                                                                            */
@@ -716,7 +716,7 @@ if ($action != 'create' && $action != 'edit')
 		{
 			if (!isset($object->no_button_edit) || $object->no_button_edit <> 1) print '<a class="butAction" href="'.$_SERVER["PHP_SELF"].'?action=edit&amp;id='.$object->id.'">'.$langs->trans("Modify").'</a>';
 
-			if (!isset($object->no_button_copy) || $object->no_button_copy <> 1)
+			/*if (!isset($object->no_button_copy) || $object->no_button_copy <> 1)
 			{
 				if (!empty($conf->use_javascript_ajax) && empty($conf->dol_use_jmobile))
 				{
@@ -724,7 +724,7 @@ if ($action != 'create' && $action != 'edit')
 				} else {
 					print '<a class="butAction" href="'.$_SERVER["PHP_SELF"].'?action=clone&amp;id='.$object->id.'">'.$langs->trans("ToClone").'</a>';
 				}
-			}
+			}*/
 		}
 
 		if ($usercandelete)
@@ -735,7 +735,7 @@ if ($action != 'create' && $action != 'edit')
 				{
 					print '<span id="action-delete" class="butActionDelete">'.$langs->trans('Delete').'</span>'."\n";
 				} else {
-					print '<a class="butActionDelete" href="'.$_SERVER["PHP_SELF"].'?action=delete&amp;token='.newToken().'&amp;id='.$object->id.'">'.$langs->trans("Delete").'</a>';
+					print '<a class="butActionDelete" onclick="return confirm(\'Voulez-vous vraiment supprimer ce passager ! \');" href="'.$_SERVER["PHP_SELF"].'?action=delete&amp;token='.newToken().'&amp;id='.$object->id.'">'.$langs->trans("Delete").'</a>';
 				}
 			} else {
 				print '<a class="butActionRefused classfortooltip" href="#" title="'.$langs->trans("PassengerIsUsed").'">'.$langs->trans("Delete").'</a>';
